@@ -1,17 +1,34 @@
 class ConfirmationsController < Devise::ConfirmationsController
+
   def show
-    self.resource = resource_class.find_by_confirmation_token(params[:confirmation_token])
+    if params[:confirmation_token].present?
+      @original_token = params[:confirmation_token]
+    elsif params[resource_name].try(:[], :confirmation_token).present?
+      @original_token = params[resource_name][:confirmation_token]
+    end
+
+    self.resource = resource_class.find_by_confirmation_token Devise.token_generator.
+      digest(self, :confirmation_token, @original_token)
+
     super if resource.nil? or resource.confirmed?
   end
 
   def confirm
-    self.resource = resource_class.find_by_confirmation_token(params[resource_name][:confirmation_token])
-    if resource.update_attributes(params[resource_name].except(:confirmation_token)) && resource.password_match?
+    @original_token = params[resource_name].try(:[], :confirmation_token)
+    self.resource = resource_class.find_by_confirmation_token! @original_token
+    resource.assign_attributes(permitted_params) unless params[resource_name].nil?
+
+    if resource.valid? && resource.password_match?
       self.resource.confirm!
       set_flash_message :notice, :confirmed
-      sign_in_and_redirect(resource_name, resource)
+      sign_in_and_redirect resource_name, resource
     else
-      render :action => "show"
+      render :action => 'show'
     end
   end
+
+ private
+   def permitted_params # Make sure to update this in the users controller as well
+     params.require(resource_name).permit(:username, :password, :password_confirmation, :remember_me, :team_id, :team_captain, :email, :name, :school, :year_in_school, :gender, :age, :area_of_study, :location, :personal_email, :resume)
+   end
 end
